Testing for the encoding/decoding method of the ChainCodeCoder Class is implemented.
In ChainCodeCoder Class the visibility of STEPS_MASK and DIRECTION_MASK  is changed from private to public to facilitate automatic testing.

The Class DecodedData has been moved to its own file and made public in order to use it within the test.

An toString Method was added/overwritten for the DecodedData Class. This is not strictly necessary, but might be handy.